+++ src/sys/os/term/backend/linux/mod.rs
@@ -4,6 +4,10 @@
 //
 
 mod term; // TermLinux
+mod poll; // impl polling
+mod cap; // impl capabilities
+
+mod buf; // (TermLinuxInputBuf)
 mod restore; // (TermLinuxRestore, TermLinuxRestoreFlags)
 
 crate::structural_mods! { // _mods, _crate_internals
@@ -11,6 +15,9 @@ crate::structural_mods! { // _mods, _crate_internals
         pub use super::term::*;
     }
     _crate_internals {
-        pub use super::restore::*;
+        pub use super::{
+            buf::*,
+            restore::*,
+        };
     }
 }